+++ src/Fauna/world.cpp
@@ -111,8 +111,11 @@ void World::simulate_day(const Date& date, const bool do_herbivores) {
       continue;
     }
 
+    // Whether herbivores shall be (re-)established today.
+    bool establish_as_needed = false;
+
     // If there was no initial establishment yet, we may do this now.
-    bool establish_if_needed = !sim_unit.is_initial_establishment_done();
+    if (!sim_unit.is_initial_establishment_done()) establish_as_needed = true;
 
     // If one check interval has passed, we will check if HFTs have died out
     // and need to be re-established.
@@ -121,17 +124,12 @@ void World::simulate_day(const Date& date, const bool do_herbivores) {
     if (days_since_last_establishment >=
             get_params().herbivore_establish_interval &&
         get_params().herbivore_establish_interval > 0) {
-      establish_if_needed = true;
+      establish_as_needed = true;
       days_since_last_establishment = 0;
     }
 
     // Keep track of the establishment cycle.
-    days_since_last_establishment++;
-
-    // Establish each HFT if it got extinct.
-    if (establish_if_needed)
-      for (auto& pop : sim_unit.get_populations())
-        if (pop->get_list().empty()) pop->establish();
+    if (do_herbivores) days_since_last_establishment++;
 
     // Create function object to delegate all simulations for this day to.
     // TODO: Create function object only once per day and for all simulation
@@ -139,7 +137,7 @@ void World::simulate_day(const Date& date, const bool do_herbivores) {
     SimulateDay simulate_day(date.get_julian_day(), sim_unit, feed_herbivores);
 
     // Call the function object.
-    simulate_day(do_herbivores);
+    simulate_day(do_herbivores, establish_as_needed);
 
     // Aggregate output.
     assert(output_aggregator.get() != NULL);